REF: move ICON_MAP to top-level widgets __init__ to avoid circular imports
