+++ superscore/widgets/page/__init__.py
@@ -1,21 +1,13 @@
-def get_page_icon_map():
+def get_page_map():
     # Don't pollute the namespace
-    from superscore.model import Collection, Readback, Setpoint, Snapshot
+    from superscore.model import Collection
     from superscore.widgets.page.entry import CollectionPage
 
     page_map = {
         Collection: CollectionPage
     }
 
-    # a list of qtawesome icon names
-    icon_map = {
-        Collection: 'mdi.file-document-multiple',
-        Snapshot: 'mdi.camera',
-        Setpoint: 'mdi.target',
-        Readback: 'mdi.book-open-variant',
-    }
-
-    return page_map, icon_map
+    return page_map
 
 
-PAGE_MAP, ICON_MAP = get_page_icon_map()
+PAGE_MAP = get_page_map()
